# Optimise depends_on chain resolution by using h5py