This reduces the time it takes to resolve `depends_on` chains by locating the target using h5py and a custom path resolver. It avoids creating intermediate `snx.Group` objects which is expensive.

For loading the 45 BIFROST analysers using 
```python
def load(fname):
    start = time.perf_counter()
    with snx.File(fname) as f:
        instrument = f['entry']['instrument']
        groups = instrument[snx.NXcrystal]
        start_load = time.perf_counter()
        crystals = {k: v[()] for k, v in groups.items()}
        end_load = time.perf_counter()
    closed = time.perf_counter()
    print(f'total: {closed-start:.3f}s   load: {end_load-start_load:.3f}s')
```
I get these results on my machine:
```
# Baseline:       total: 1.205s   load: 1.082s
# Custom resolve: total: 0.345s   load: 0.306s
```